openpose/requirements.txt
argparse
matplotlib
scipy
tqdm
requests
fire
dill
Flask
flask-cors
pillow
gevent
flask-socketio
git+https://github.com/ppwwyyxx/tensorpack.git